Collapse http telemetry up into connector
