+++ apollo-router/src/plugins/telemetry/config_new/connector.rs
@@ -1,5 +1,12 @@
 //! Connectors telemetry.
 
+use crate::services::http::HttpRequest;
+use crate::services::http::HttpResponse;
+
+pub(crate) mod attributes;
 pub(crate) mod events;
-pub(crate) mod http;
 pub(crate) mod instruments;
+pub(crate) mod selectors;
+
+pub(crate) type ConnectorRequest = HttpRequest;
+pub(crate) type ConnectorResponse = HttpResponse;
