--- apollo-router/src/plugins/telemetry/config_new/connector/http.rs
@@ -1,6 +1,0 @@
-//! Telemetry related to HTTP requests and responses for connectors.
-
-pub(crate) mod attributes;
-pub(crate) mod events;
-pub(crate) mod instruments;
-pub(crate) mod selectors;
