+++ apollo-router/src/plugins/telemetry/config_new/fixtures/connector/custom_counter_with_conditions/router.yaml
@@ -3,20 +3,19 @@ telemetry:
     instruments:
       default_requirement_level: none
       connector:
-        http:
-          not.found.count:
-            description: "Count of 404 responses from the user API"
-            type: counter
-            unit: count
-            value: unit
-            attributes:
-              "url_template":
-                connector_url_template: true
-            condition:
-              all:
-                - eq:
-                    - 404
-                    - connector_http_response_status: code
-                - eq:
-                    - "user_api"
-                    - connector_source: name
+        not.found.count:
+          description: "Count of 404 responses from the user API"
+          type: counter
+          unit: count
+          value: unit
+          attributes:
+            "url_template":
+              connector_url_template: true
+          condition:
+            all:
+              - eq:
+                  - 404
+                  - connector_http_response_status: code
+              - eq:
+                  - "user_api"
+                  - connector_source: name
