+++ apollo-router/src/plugins/telemetry/config_new/fixtures/connector/custom_histogram/router.yaml
@@ -3,18 +3,17 @@ telemetry:
     instruments:
       default_requirement_level: none
       connector:
-        http:
-          rate.limit:
-            value:
-              connector_http_response_header: "x-ratelimit-remaining"
-            unit: count
-            type: histogram
-            description: "Rate limit remaining"
-            condition:
-              all:
-                - eq:
-                  - 200
-                  - connector_http_response_status: code
-                - eq:
-                  - "user_api"
-                  - connector_source: name
+        rate.limit:
+          value:
+            connector_http_response_header: "x-ratelimit-remaining"
+          unit: count
+          type: histogram
+          description: "Rate limit remaining"
+          condition:
+            all:
+              - eq:
+                - 200
+                - connector_http_response_status: code
+              - eq:
+                - "user_api"
+                - connector_source: name
