+++ apollo-router/src/plugins/telemetry/mod.rs
@@ -10,7 +10,7 @@ use ::tracing::info_span;
 use ::tracing::Span;
 use axum::headers::HeaderName;
 use config_new::cache::CacheInstruments;
-use config_new::connector::http::instruments::ConnectorHttpInstruments;
+use config_new::connector::instruments::ConnectorInstruments;
 use config_new::instruments::InstrumentsConfig;
 use config_new::instruments::StaticInstrument;
 use config_new::Selectors;
@@ -102,7 +102,7 @@ use crate::plugins::telemetry::apollo_exporter::proto::reports::StatsContext;
 use crate::plugins::telemetry::config::AttributeValue;
 use crate::plugins::telemetry::config::MetricsCommon;
 use crate::plugins::telemetry::config::TracingCommon;
-use crate::plugins::telemetry::config_new::connector::http::events::ConnectorHttpEvents;
+use crate::plugins::telemetry::config_new::connector::events::ConnectorEvents;
 use crate::plugins::telemetry::config_new::cost::add_cost_attributes;
 use crate::plugins::telemetry::config_new::graphql::GraphQLInstruments;
 use crate::plugins::telemetry::config_new::instruments::SupergraphInstruments;
@@ -138,6 +138,7 @@ use crate::query_planner::OperationKind;
 use crate::router_factory::Endpoint;
 use crate::services::connector_service::CONNECTOR_INFO_CONTEXT_KEY;
 use crate::services::execution;
+use crate::services::http::HttpRequest;
 use crate::services::router;
 use crate::services::subgraph;
 use crate::services::subgraph::Request;
@@ -867,7 +868,7 @@ impl PluginPrivate for Telemetry {
         let static_connector_instruments = self.connector_custom_instruments.read().clone();
         ServiceBuilder::new()
             .map_future_with_request_data(
-                move |http_request: &crate::services::http::HttpRequest| {
+                move |http_request: &HttpRequest| {
                     if http_request
                         .context
                         .contains_key(CONNECTOR_INFO_CONTEXT_KEY)
@@ -877,10 +878,8 @@ impl PluginPrivate for Telemetry {
                             .instruments
                             .new_connector_instruments(static_connector_instruments.clone());
                         custom_instruments.on_request(http_request);
-                        let custom_events = req_fn_config
-                            .instrumentation
-                            .events
-                            .new_connector_http_events();
+                        let custom_events =
+                            req_fn_config.instrumentation.events.new_connector_events();
                         custom_events.on_request(http_request);
                         (
                             http_request.context.clone(),
@@ -892,27 +891,29 @@ impl PluginPrivate for Telemetry {
                 },
                 move |(context, custom_telemetry): (
                     Context,
-                    Option<(ConnectorHttpInstruments, ConnectorHttpEvents)>,
+                    Option<(ConnectorInstruments, ConnectorEvents)>,
                 ),
                       f: BoxFuture<
                     'static,
                     Result<crate::services::http::HttpResponse, BoxError>,
                 >| {
                     async move {
-                        let result = f.await;
                         if let Some((custom_instruments, custom_events)) = custom_telemetry {
+                            let result = f.await;
                             match &result {
-                                Ok(resp) => {
-                                    custom_instruments.on_response(resp);
-                                    custom_events.on_response(resp);
+                                Ok(http_response) => {
+                                    custom_instruments.on_response(http_response);
+                                    custom_events.on_response(http_response);
                                 }
                                 Err(err) => {
                                     custom_instruments.on_error(err, &context);
                                     custom_events.on_error(err, &context);
                                 }
                             }
+                            result
+                        } else {
+                            f.await
                         }
-                        result
                     }
                 },
             )
