+++ docs/source/configuration/telemetry/instrumentation/events.mdx
@@ -232,7 +232,6 @@ telemetry:
       subgraph:
           # Custom event configuration for subgraph service ...
       connector:
-        http:
           # Custom event configuration for HTTP connectors ...
 ```
